Add option to pass 'outsideDeactivate' Element
This needs to be brought back since there is several cases where we want to keep items selected when they are clicking other places on the page.

The best case for this is MJ where a user selects a file and then clicks something in the ribbon it then deactivates since its listening for all clicks on the body.

Signed-off-by: Austin McDaniel <amcdaniel2@gmail.com>